+++ src/main/flexisip-conference.cc
@@ -265,8 +265,12 @@ static void depthFirstSearch(string& path, const GenericEntry* config, list<stri
 	}
 }
 
-static void dump_config(
-    ConfigManager& cfg, const string& dump_cfg_part, bool with_experimental, bool dumpDefault, const string& format) {
+static void dump_config(ConfigManager& cfg,
+                        const string& dump_cfg_part,
+                        bool with_experimental,
+                        bool dumpDefault,
+                        bool removeDeprecated,
+                        const string& format) {
 	cfg.applyOverrides(true);
 
 	auto* rootStruct = cfg.getEditableRoot();
@@ -310,6 +314,7 @@ static void dump_config(
 		throw ExitFailure{"invalid output format '" + format + "'"};
 	}
 	dumper->setDumpExperimentalEnabled(with_experimental);
+	dumper->setRemoveDeprecated(removeDeprecated);
 	dumper->dump(cout);
 }
 
@@ -408,6 +413,10 @@ int flexisip_conference::main(int argc, const char* argv[]) {
                                                                        "to the default value and the default value has "
                                                                        "changed.",
                                                                        cmd);
+	TCLAP::SwitchArg         removeDeprecated("", "remove-deprecated", "When dumping the configuration, remove all "
+                                                                       "deprecated parameters even if they had a value "
+                                                                       "explicitly set.",
+                                                                       cmd);
 	TCLAP::SwitchArg              dumpMibs("",  "dump-mibs", 		   "Dump the MIB files for Flexisip performance "
                                                                        "counters and other related SNMP items in the "
                                                                        "standard output.",
@@ -478,7 +487,7 @@ int flexisip_conference::main(int argc, const char* argv[]) {
 	cfg->setOverrideMap(oset);
 
 	if (const auto module = dumpAll ? "all" : dumpDefault.getValue(); !module.empty()) {
-		dump_config(*cfg, module, displayExperimental, true, dumpFormat.getValue());
+		dump_config(*cfg, module, displayExperimental, true, removeDeprecated, dumpFormat.getValue());
 		return EXIT_SUCCESS;
 	}
 	auto* rootCfg = cfg->getEditableRoot();
@@ -526,7 +535,7 @@ int flexisip_conference::main(int argc, const char* argv[]) {
 	}
 
 	if (rewriteConf) {
-		dump_config(*cfg, "all", displayExperimental, false, "file");
+		dump_config(*cfg, "all", displayExperimental, false, removeDeprecated, "file");
 		return EXIT_SUCCESS;
 	}